+++ src/list.js
@@ -1,17 +1,13 @@
 /**
+ * This file contains the extensions for core/list.
+ *
+ * @package nested-ordered-lists-block-editor
+ */
+
+/**
  * Import language-support.
  */
 const { __ } = wp.i18n;
-
-/**
- * Define which blocks will be changed by the following functions.
- * Only core/list will be changed.
- *
- * @type {string[]}
- */
-const enableSidebarSelectOnBlocks = [
-    'core/list'
-];
 
 /**
  * Set more imports and constants.
@@ -21,7 +17,7 @@ import {
     __experimentalNumberControl as NumberControl,
 } from '@wordpress/components';
 const { createHigherOrderComponent } = wp.compose;
-const { Fragment } = wp.element;
+const { Fragment, useEffect } = wp.element;
 const {
     InspectorControls,
     BlockControls
@@ -30,9 +26,13 @@ const {
     ToolbarGroup,
     ToolbarButton,
     PanelBody,
-    CheckboxControl
+    CheckboxControl,
+	ToolbarDropdownMenu,
 } = wp.components;
+import { more } from '@wordpress/icons';
 import { list_max_level } from './config';
+import { enableSidebarSelectOnBlocks, removeIconsFromChildren } from './helper';
+import { useSelect } from '@wordpress/data';
 import icons from './icons';
 
 /**
@@ -192,6 +192,7 @@ const addListAttributes = ( settings, name ) => {
         return settings;
     }
 
+	// add the new attributes.
     settings.attributes = Object.assign(settings.attributes, {
         nestedList: { type: 'boolean', default: false },
         listIntent: { type: 'boolean', default: false },
@@ -201,6 +202,7 @@ const addListAttributes = ( settings, name ) => {
 		clientId: { type: 'string', default: false }
     });
 
+	// return resulting settings for this list.
     return settings;
 };
 wp.hooks.addFilter(
@@ -210,68 +212,10 @@ wp.hooks.addFilter(
 );
 
 /**
- * Define additional block-elements for output in Block Editor.
+ * Add our custom options for the list block.
  */
-function setAttributesInEditor( BlockListBlock ) {
+const addOptionsToListBlock = createHigherOrderComponent( ( BlockEdit ) => {
     return ( props ) => {
-
-        // If current block is not allowed.
-        if ( ! enableSidebarSelectOnBlocks.includes( props.name ) ) {
-            return <BlockListBlock { ...props } />;
-        }
-
-        /**
-         * Get attributes
-         */
-        const { nestedList, listIntent } = props.attributes;
-
-        // collect classes for list.
-        let nolgClassName = '';
-
-        if( props.attributes.type ) {
-            // set class name.
-            nolgClassName = nolgClassName + ' nolg-style';
-        }
-
-        /**
-         * If nested list is activated for this Block add our own classes to it for styling.
-         */
-        if ( nestedList ) {
-            // set class name.
-            nolgClassName = nolgClassName + ' nolg-list';
-
-			// get parent settings for nested list, if available and enabled on parent.
-			const parentBlocks = wp.data.select( 'core/block-editor' ).getBlockParentsByBlockName(props.clientId, enableSidebarSelectOnBlocks );
-			const parentAttributes = wp.data.select('core/block-editor').getBlocksByClientId(parentBlocks);
-			if( parentAttributes[0] && parentAttributes[0].attributes.nestedList && parentAttributes[0].attributes.inheritSettings ) {
-				props.attributes.ordered = parentAttributes[0].attributes.ordered;
-				props.attributes.type = parentAttributes[0].attributes.type;
-			}
-
-            // get counter-reset-styling.
-            let counterReset = getCounterReset({...props.wrapperProps}, props.attributes)
-
-            if( listIntent ) {
-                return <BlockListBlock { ...props } wrapperProps={counterReset} className={ nolgClassName + ' nolg-list-intent' } />
-            }
-            return <BlockListBlock { ...props } wrapperProps={counterReset} className={nolgClassName} />
-        }
-
-        return <BlockListBlock { ...props } className={nolgClassName} />
-    };
-}
-wp.hooks.addFilter(
-    'editor.BlockListBlock',
-    'nolg/set-attributes-in-editor',
-    setAttributesInEditor
-);
-
-/**
- * Add our custom options in sidebar.
- */
-const addOptionInSidebar = createHigherOrderComponent( ( BlockEdit ) => {
-    return ( props ) => {
-
         // If current block is not allowed.
         if ( ! enableSidebarSelectOnBlocks.includes( props.name ) ) {
             return (
@@ -284,6 +228,27 @@ const addOptionInSidebar = createHigherOrderComponent( ( BlockEdit ) => {
         let { nestedList, listIntent, inheritSettings } = attributes;
 		attributes.clientId = props.clientId;
 
+		// get the iconsets.
+		const iconsets = useSelect((select) => {
+			return select('core').getEntityRecords( 'taxonomy', 'nolg_icon_set' );
+		});
+
+		// generate iconset list for options.
+		let iconsets_to_use = [];
+		let active_iconset = false;
+		if( iconsets !== null ) {
+			iconsets.map((iconset) => {
+				if( iconset.slug === attributes.type ) {
+					active_iconset = iconset;
+				}
+				iconsets_to_use.push({
+					'title': iconset.name,
+					'icon': iconset.slug,
+					'onClick': () => onClickAttributeType(iconset.slug)
+				})
+			});
+		}
+
         // get parent settings for nested list, if available and enabled on parent.
         const parentBlocks = wp.data.select( 'core/block-editor' ).getBlockParentsByBlockName(props.clientId, enableSidebarSelectOnBlocks);
         const parentAttributes = wp.data.select('core/block-editor').getBlocksByClientId(parentBlocks);
@@ -291,18 +256,19 @@ const addOptionInSidebar = createHigherOrderComponent( ( BlockEdit ) => {
         if( parentAttributes[0] && parentAttributes[0].attributes.nestedList && parentAttributes[0].attributes.inheritSettings ) {
             inheritedSettings = true;
             attributes.ordered = parentAttributes[0].attributes.ordered;
-            attributes.type = parentAttributes[0].type;
-            nestedList = true;
+            attributes.type = parentAttributes[0].attributes.type;
+            nestedList = parentAttributes[0].attributes.nestedList;
         }
 
         /**
-         * Update type of list.
+         * Update type attribute of list.
          *
          * @param value
          */
         const onClickAttributeType = ( value ) => {
             {
                 {
+					// set the new setting.
                     attributes.type !== value && setAttributes({
                         type: value,
                         ordered: true,
@@ -310,14 +276,21 @@ const addOptionInSidebar = createHigherOrderComponent( ( BlockEdit ) => {
                     });
                 }
                 {
+					// remove the type setting to disable our custom list functions.
                     attributes.type === value && setAttributes({
-                        type: ''
+                        type: '',
+						className: ''
                     });
                 }
+
+				// update all children to remove their custom settings used for the previous type.
+				removeIconsFromChildren( props.clientId, attributes.type === value ? undefined : value );
             }
         }
 
-        // create output
+		console.log(props, BlockEdit)
+
+        // create output.
         return (
             <Fragment>
                 <BlockEdit { ...props } />
@@ -347,7 +320,18 @@ const addOptionInSidebar = createHigherOrderComponent( ( BlockEdit ) => {
                                 isActive={attributes.type === 'i2'}
                                 onClick={ value => onClickAttributeType('i2') }
                             />
+							{active_iconset && <ToolbarButton
+								icon={ active_iconset.slug }
+								label={ active_iconset.name }
+								isActive={true}
+								onClick={ value => onClickAttributeType(active_iconset.slug) }
+							/>}
                         </ToolbarGroup>
+						<ToolbarDropdownMenu
+							icon={ more }
+							label={ __( 'Select an iconset', 'nested-ordered-lists-for-block-editor' ) }
+							controls={ iconsets_to_use }
+						/>
                     </BlockControls>
                 }
                 {attributes.ordered && !inheritedSettings && <InspectorControls>
@@ -401,34 +385,84 @@ const addOptionInSidebar = createHigherOrderComponent( ( BlockEdit ) => {
 wp.hooks.addFilter(
     'editor.BlockEdit',
     'nolg/add-option-in-sidebar',
-    addOptionInSidebar
+	addOptionsToListBlock
+);
+
+/**
+ * Define additional block-attributes for output in Block Editor.
+ */
+function setAttributesInEditor( BlockListBlock ) {
+	return ( props ) => {
+
+		// If current block is not allowed.
+		if ( ! enableSidebarSelectOnBlocks.includes( props.name ) ) {
+			return <BlockListBlock { ...props } />;
+		}
+
+		// get parent settings for nested list, if available and enabled on child.
+		const parentBlocks = wp.data.select( 'core/block-editor' ).getBlockParentsByBlockName(props.clientId, enableSidebarSelectOnBlocks );
+		const parentAttributes = wp.data.select('core/block-editor').getBlocksByClientId(parentBlocks);
+		if( parentAttributes[0] && parentAttributes[0].attributes.nestedList && parentAttributes[0].attributes.inheritSettings ) {
+			useEffect(() => {
+				props.setAttributes( {
+					nestedList: parentAttributes[0].attributes.nestedList,
+					ordered: parentAttributes[0].attributes.ordered,
+					type: parentAttributes[0].attributes.type
+				} );
+			}, []);
+		}
+
+		// get attributes.
+		const { ordered, nestedList, listIntent, type } = props.attributes;
+
+		// bail if list is not ordered.
+		if( ! ordered ) {
+			return <BlockListBlock { ...props } />;
+		}
+
+		// bail if list is not nested.
+		if( ! nestedList ) {
+			return <BlockListBlock { ...props } />;
+		}
+
+		// bail if type is unknown.
+		if( type === undefined ) {
+			return <BlockListBlock { ...props } />;
+		}
+
+		// bail if no type is set.
+		if( type.length === 0 ) {
+			return <BlockListBlock { ...props } />;
+		}
+
+		// collect classes for list.
+		let nolgClassName = 'nolg-style nolg-list';
+
+		// set class to intended list.
+		if( listIntent ) {
+			nolgClassName = nolgClassName + ' nolg-list-intent';
+		}
+
+		// get counter-reset-styling.
+		let counterReset = getCounterReset({...props.wrapperProps}, props.attributes)
+
+		// return resulting block.
+		return <BlockListBlock { ...props } wrapperProps={counterReset} className={nolgClassName} />
+	};
+}
+wp.hooks.addFilter(
+	'editor.BlockListBlock',
+	'nolg/set-attributes-in-editor',
+	setAttributesInEditor
 );
 
 /**
  * Save custom attribute for output in frontend.
  */
 const saveAttributesForFrontend = ( extraProps, blockType, attributes ) => {
-    // Do nothing if it's another block than our defined ones.
+    // bail if block is not core/list.
     if ( !enableSidebarSelectOnBlocks.includes( blockType.name ) ) {
         return extraProps;
-    }
-
-    // Do also nothing if it is not ordered.
-    if( !attributes.ordered ) {
-        return extraProps;
-    }
-
-    // get attributes.
-    const { nestedList, listIntent } = attributes;
-
-    // add class if we use a list-style.
-    if( attributes.type ) {
-        extraProps.className = classnames( extraProps.className, 'nolg-style' );
-    }
-
-    // change some settings if our nested list is active.
-    if( ! nestedList ) {
-		return extraProps;
 	}
 
 	// get parent settings for nested list, if available and enabled on parent.
@@ -437,11 +471,36 @@ const saveAttributesForFrontend = ( extraProps, blockType, attributes ) => {
 	if( parentAttributes[0] && parentAttributes[0].attributes.nestedList && parentAttributes[0].attributes.inheritSettings ) {
 		attributes.ordered = parentAttributes[0].attributes.ordered;
 		attributes.type = parentAttributes[0].attributes.type;
-		extraProps.className = classnames( extraProps.className, 'nolg-style' );
 	}
 
-	// add our own class.
-	extraProps.className = classnames( extraProps.className, 'nolg-list' );
+	// get attributes.
+	const { ordered, nestedList, listIntent, type } = attributes;
+
+    // bail if list is not ordered.
+    if( ! ordered ) {
+        return extraProps;
+    }
+
+	// bail if this is not a nested list.
+	if( ! nestedList ) {
+		return extraProps;
+	}
+
+	// bail if type is unknown.
+	if( type === undefined ) {
+		return extraProps;
+	}
+
+	// bail if type is empty.
+	if( type.length === 0 ) {
+		return extraProps;
+	}
+
+	// add our style.
+	extraProps.className = classnames(extraProps.className, 'nolg-style');
+	extraProps.className = classnames(extraProps.className, 'nolg-list');
+
+	// add our own class if list is intended.
 	if( listIntent ) {
 		// add additional class if intent should be used.
 		extraProps.className = classnames(extraProps.className, 'nolg-list-intent')
@@ -455,3 +514,4 @@ wp.hooks.addFilter(
     'nolg/save-attributes-for-frontend',
     saveAttributesForFrontend
 );
+